+++ src/main/java/com/pawwithu/connectdog/domain/auth/controller/AuthController.java
@@ -1,7 +1,14 @@
 package com.pawwithu.connectdog.domain.auth.controller;
 
+import com.pawwithu.connectdog.domain.auth.dto.request.EmailRequest;
+import com.pawwithu.connectdog.domain.auth.dto.request.IntermediaryPhoneRequest;
 import com.pawwithu.connectdog.domain.auth.dto.request.RefreshTokenRequest;
+import com.pawwithu.connectdog.domain.auth.dto.request.VolunteerPhoneRequest;
+import com.pawwithu.connectdog.domain.auth.dto.response.EmailResponse;
+import com.pawwithu.connectdog.domain.auth.dto.response.IntermediaryEmailResponse;
+import com.pawwithu.connectdog.domain.auth.dto.response.VolunteerEmailResponse;
 import com.pawwithu.connectdog.domain.auth.service.AuthService;
+import com.pawwithu.connectdog.domain.auth.service.EmailService;
 import com.pawwithu.connectdog.domain.oauth.dto.response.LoginResponse;
 import com.pawwithu.connectdog.error.dto.ErrorResponse;
 import com.pawwithu.connectdog.jwt.service.JwtService;
@@ -29,6 +36,7 @@ public class AuthController {
 
     private final JwtService jwtService;
     private final AuthService authService;
+    private final EmailService emailService;
 
     @Operation(summary = "토큰 재발행", description = "AccessToken, RefreshToken 재발행 합니다.",
             responses = {@ApiResponse(responseCode = "200", description = "토큰 재발행 성공")
@@ -94,6 +102,42 @@ public class AuthController {
     public ResponseEntity<Void> intermediariesWithdraw(HttpServletRequest request, @AuthenticationPrincipal UserDetails loginUser) {
         authService.intermediariesWithdraw(request, loginUser.getUsername());
         return ResponseEntity.noContent().build();
+    }
+
+    @Operation(summary = "이메일 찾기 - 봉사자 휴대폰 번호로 이메일 찾기", description = "봉사자 휴대폰 번호로 이메일을 찾습니다.",
+            responses = {@ApiResponse(responseCode = "200", description = "봉사자 휴대폰 번호로 이메일 찾기 성공")
+                    , @ApiResponse(responseCode = "400"
+                    , description = "V1, 휴대폰 번호는 필수 입력 값입니다. \t\n M1, 해당 이동봉사자를 찾을 수 없습니다."
+                    , content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+            })
+    @PostMapping("/volunteers/search/email")
+    public ResponseEntity<VolunteerEmailResponse> findVolunteerEmail(@RequestBody @Valid VolunteerPhoneRequest request) {
+        VolunteerEmailResponse response = authService.findVolunteerEmail(request);
+        return ResponseEntity.ok(response);
+    }
+
+    @Operation(summary = "이메일 찾기 - 모집자 휴대폰 번호로 이메일 찾기", description = "모집자 휴대폰 번호로 이메일을 찾습니다.",
+            responses = {@ApiResponse(responseCode = "200", description = "모집자 휴대폰 번호로 이메일 찾기 성공")
+                    , @ApiResponse(responseCode = "400"
+                    , description = "V1, 휴대폰 번호는 필수 입력 값입니다. \t\n M2, 해당 이동봉사 중개를 찾을 수 없습니다."
+                    , content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+            })
+    @PostMapping("/intermediaries/search/email")
+    public ResponseEntity<IntermediaryEmailResponse> findIntermediaryEmail(@RequestBody @Valid IntermediaryPhoneRequest request) {
+        IntermediaryEmailResponse response = authService.findIntermediaryEmail(request);
+        return ResponseEntity.ok(response);
+    }
+
+    @Operation(summary = "비밀번호 찾기 - 이메일 인증번호 전송", description = "입력한 이메일로 인증번호를 전송합니다.",
+            responses = {@ApiResponse(responseCode = "200", description = "이메일 인증번호 전송 성공")
+                    , @ApiResponse(responseCode = "400"
+                    , description = "V1, 이메일 형식에 맞지 않습니다. \t\n V1, 이메일은 필수 입력 값입니다. \t\n A1, 이미 존재하는 이메일입니다. \t\n A4, 이메일 인증 코드 전송을 실패했습니다."
+                    , content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+            })
+    @PostMapping(value = {"/volunteers/search/send-email", "/intermediaries/search/send-email"})
+    public ResponseEntity<EmailResponse> sendEmail(@RequestBody @Valid EmailRequest request){
+        EmailResponse emailResponse = emailService.sendEmailWithoutAuth(request);
+        return ResponseEntity.ok(emailResponse);
     }
 
 }
